renamed files under the banner deploy directory to make them shorter and easier to read.
